Introduce a unit test for the Recorder
Basic coverage on writing and reading messages,
and an assertion related to how the messages
are joined together.

This is a WIP PR. Just to get the dependencies
working, I ended up including all of rails, and
the Redis gem. It is likely possible to load
less code than this.

Also, I’m wondering whether the intention is
to use the Redis gem or a mock or fake redis
object for the purposes of unit tests against
the Recorder internals. This PR includes
the Redis gem to satisfy the dependency.

The mattr_accessor redis also was nil,
which caused Redis.new() to fail initialization.
In the test, I passed it a hash of a minimal
amount of configuration, which would be like a 
local Redis instance, which allowed it to 
init properly.